Version 2.0.0(268b5e6)
1. update boot.bin to v1.1.1;
2. generated bin name changed, irom1.bin to drom0.bin;
3. add new map mode;
4. add WPS support;
5. add WPA2-Enterprise support;
6. add AMPDU support;
7. add esptool.py;
8. update gen_appbin.py;
9. include folder structure change;
10.support some makefile options;
11.add esp.conf in project_template;
12.add lwip source code;
13.add cjson source code;
14.update README.md;
15.ssl bug fix;
16.upgrade bug fix;
17.many other bugs fix, many other changes;

+++ tools/gen_appbin.py
@@ -30,6 +30,7 @@ import zlib
 
 
 TEXT_ADDRESS = 0x40040000
+TEXT_ADDRESS_APPCPU = 0x3FFA8000
 # app_entry = 0
 # data_address = 0x3ffb0000
 # data_end  = 0x40000000
@@ -111,21 +112,19 @@ def getFileCRC(_path):
         return 0 
     return crc
 
-def gen_appbin():
+def gen_appbin(app_cpu_flag,combine_bin_flag):
     global chk_sum
     global crc_sum
     global blocks 
     global irom0_addr
     global irom1_addr
 
-    if len(sys.argv) != 3:
-        print 'Usage: gen_appbin.py eagle.app.out'
-        sys.exit(0)
-
     elf_file = sys.argv[1]
-    
-    ld_fd = open(sys.argv[2] + "/eagle.pro.v7.ld")
+    app_bin_file = sys.argv[5]
+    ld_fd = open(sys.argv[2])
     ld_data = ld_fd.read()
+    app_cpu_flag = int(app_cpu_flag)
+    combine_bin_flag = int(combine_bin_flag)
     match = re.search("irom0_0_seg.+?org = (0x[0-9a-fA-F]+),",ld_data)
     if match is not None:
         irom0_addr = match.group(1)
@@ -136,31 +135,32 @@ def gen_appbin():
         irom0_len = match.group(1)
         irom0_len = int(irom0_len,base=16)
         #print irom0_len
-    match = re.search("irom0_1_seg.+?org = (0x[0-9a-fA-F]+),",ld_data)
+    match = re.search("drom0_0_seg.+?org = (0x[0-9a-fA-F]+),",ld_data)
     if match is not None:
-        irom1_addr = match.group(1)
-        #irom1_addr = int(addr,base=16)
-        #print irom1_addr
-    match = re.search("irom0_1_seg.+?len = (0x[0-9a-fA-F]+)",ld_data)
+        drom0_addr = match.group(1)
+        #drom0_addr = int(addr,base=16)
+        #print drom0_addr
+    match = re.search("drom0_0_seg.+?len = (0x[0-9a-fA-F]+)",ld_data)
     if match is not None:
-        irom1_len = match.group(1)
-        irom1_len = int(irom1_len,base=16)
+        drom0_len = match.group(1)
+        drom0_len = int(drom0_len,base=16)
         #print irom1_len
     flash_data_line  = 16
     data_line_bits = 0xf
 
     irom0text_bin_name = 'eagle.app.v7.irom0text.bin'
-    irom1text_bin_name = 'eagle.app.v7.irom1text.bin'
+    drom0text_bin_name = 'eagle.app.v7.drom0text.bin'
     text_bin_name = 'eagle.app.v7.text.bin'
     data_bin_name = 'eagle.app.v7.data.bin'
     rodata_bin_name = 'eagle.app.v7.rodata.bin'
-    flash_bin_name ='eagle.app.flash.bin'
-    Dcache_bin_name = 'irom1.bin'
-    Icache_flash_bin_name = 'irom0_flash.bin'
+    flash_bin_name = sys.argv[6] + '/user.ota'
+    Dcache_bin_name = sys.argv[6] + '/drom0.bin'
+    Icache_flash_bin_name = sys.argv[6] + '/irom0_flash.bin'
 
     BIN_MAGIC_FLASH  = 0xE9
     BIN_MAGIC_IROM0  = 0xEA
-    BIN_MAGIC_IROM1  = 0xEC
+    BIN_MAGIC_DROM0  = 0xEC
+    BIN_MAGIC_APPCPU = 0xE1
     data_str = ''
     sum_size = 0
 
@@ -196,8 +196,18 @@ def gen_appbin():
     for line in lines:
         m = p.search(line)
         if m != None:
-            data_start_addr = m.group(1)
-                        
+            data_start_addr = m.group(1)        
+            if app_cpu_flag != 1:
+                data_start_addr = m.group(1)
+                data_start_addr = long(data_start_addr,16)
+            elif app_cpu_flag == 1 and long(data_start_addr,16) >=0x3FF80000 and long(data_start_addr,16) <= 0x3FF97FFF: 
+                data_start_addr = long(data_start_addr,16) + 0x28000
+            elif app_cpu_flag == 1 and long(data_start_addr,16) >=0x3FF98000 and long(data_start_addr,16) <= 0x3FFBC000: 
+                data_start_addr = long(data_start_addr,16) - 0x18000
+            else:               
+                data_start_addr = m.group(1)
+                data_start_addr = long(data_start_addr,16)
+                                        
 
     rodata_start_addr = '0'
     p = re.compile('(\w*)(\sA\s)(_rodata_start)$')
@@ -205,6 +215,16 @@ def gen_appbin():
         m = p.search(line)
         if m != None:
             rodata_start_addr = m.group(1)
+            if app_cpu_flag != 1:
+                rodata_start_addr = m.group(1)
+                rodata_start_addr = long(rodata_start_addr,16)
+            elif app_cpu_flag == 1 and long(rodata_start_addr,16) >=0x3FF80000 and long(rodata_start_addr,16) <= 0x3FF97FFF: 
+                rodata_start_addr = long(rodata_start_addr,16) + 0x28000
+            elif app_cpu_flag == 1 and long(rodata_start_addr,16) >=0x3FF98000 and long(rodata_start_addr,16) <= 0x3FFBC000: 
+                rodata_start_addr = long(rodata_start_addr,16) - 0x18000
+            else:               
+                rodata_start_addr = m.group(1)
+                rodata_start_addr = long(rodata_start_addr,16)
                         
     if os.path.getsize(text_bin_name):
         blocks = blocks + 1
@@ -212,54 +232,59 @@ def gen_appbin():
         blocks = blocks + 1
     if os.path.getsize(rodata_bin_name):
         blocks = blocks + 1
-            
-    fp1 = open(irom1text_bin_name,'rb')
-    fp1.seek(0,os.SEEK_END)
-    Dcache_data_len = fp1.tell()
-    # print data_len
-    if Dcache_data_len :       
-        Dcache_data_len = struct.pack('<BBBBI',BIN_MAGIC_IROM1,1,0,0,long(irom1_addr,16))
-        sum_size = len(Dcache_data_len)
-        write_file(flash_bin_name,Dcache_data_len)
-        #write_file(Dcache_bin_name,data_bin)
-        # irom1.text.bin
-        combine_bin(irom1text_bin_name,flash_bin_name,irom1_len,0)
-        open(Dcache_bin_name, "wb").write(open(flash_bin_name, "rb").read())
+    if (app_cpu_flag != 1):
+        fp1 = open(drom0text_bin_name,'rb')
+        fp1.seek(0,os.SEEK_END)
+        Dcache_data_len = fp1.tell()
+        if Dcache_data_len:       
+            Dcache_data_len = struct.pack('<BBBBI',BIN_MAGIC_DROM0,1,0,0,long(drom0_addr,16))
+            sum_size = len(Dcache_data_len)
+            write_file(flash_bin_name,Dcache_data_len)
+            #write_file(Dcache_bin_name,data_bin)
+            # drom0.text.bin
+            combine_bin(drom0text_bin_name,flash_bin_name,drom0_len,0)
+            open(Dcache_bin_name, "wb").write(open(flash_bin_name, "rb").read())
         
     # write irom0 bin head
-    fp2 = open(irom0text_bin_name,'rb')
-    fp2.seek(0,os.SEEK_END)
-    data_len = fp2.tell()
+        fp2 = open(irom0text_bin_name,'rb')
+        fp2.seek(0,os.SEEK_END)
+        data_len = fp2.tell()
     # print data_len
-    if data_len :  
-        data_bin = struct.pack('<BBBBI',BIN_MAGIC_IROM0,blocks + 1,0,0,long(irom0_addr,16)) 
-        sum_size = len(data_bin)
-        write_file(flash_bin_name,data_bin)
-        #write_file(Icache_flash_bin_name,data_bin)
-        # irom0.text.bin
-        icache_flash_size = data_len +  os.path.getsize(text_bin_name) + os.path.getsize(data_bin_name) + os.path.getsize(rodata_bin_name) + 36
-        combine_bin(irom0text_bin_name,flash_bin_name,icache_flash_size,0)
-        #combine_bin(irom0text_bin_name,Icache_flash_bin_name,0x0,0)     
+        if data_len :  
+            data_bin = struct.pack('<BBBBI',BIN_MAGIC_IROM0,blocks + 1,0,0,long(irom0_addr,16)) 
+            sum_size = len(data_bin)
+            write_file(flash_bin_name,data_bin)
+
+            # irom0.text.bin
+            if combine_bin_flag != 1:
+                icache_flash_size = data_len +  os.path.getsize(text_bin_name) + os.path.getsize(data_bin_name) + os.path.getsize(rodata_bin_name) + 60 + 8 * blocks
+            else :
+                icache_flash_size = data_len +  os.path.getsize(text_bin_name) + os.path.getsize(data_bin_name) + os.path.getsize(rodata_bin_name) + 60 + 8 * blocks + os.path.getsize(app_bin_file)
+
+            combine_bin(irom0text_bin_name,flash_bin_name,icache_flash_size,0)  
 
     # text.bin
     
-    
-    data_bin = struct.pack('<BBBBI',BIN_MAGIC_FLASH,blocks,0,0,long(entry_addr,16))
+    if app_cpu_flag != 1:  
+        data_bin = struct.pack('<BBBBI',BIN_MAGIC_FLASH,blocks,0,0,long(entry_addr,16))
+    else:
+        data_bin = struct.pack('<BBBBI',BIN_MAGIC_APPCPU,blocks,0,0,long(entry_addr,16))
     sum_size = len(data_bin)
     write_file(flash_bin_name,data_bin)
-    #write_file(Icache_flash_bin_name,data_bin)
-    
-    combine_bin(text_bin_name,flash_bin_name,TEXT_ADDRESS,1)
-    #combine_bin(text_bin_name,Icache_flash_bin_name,TEXT_ADDRESS,0)
+
+    if app_cpu_flag != 1:
+        combine_bin(text_bin_name,flash_bin_name,TEXT_ADDRESS,1)
+    else: 
+        combine_bin(text_bin_name,flash_bin_name,TEXT_ADDRESS_APPCPU,1)
+
 
     # data.bin
     if data_start_addr:
-        combine_bin(data_bin_name,flash_bin_name,long(data_start_addr,16),1)
-        #combine_bin(data_bin_name,Icache_flash_bin_name,long(data_start_addr,16),0)
+        combine_bin(data_bin_name,flash_bin_name,data_start_addr,1)
 
     # rodata.bin
-    combine_bin(rodata_bin_name,flash_bin_name,long(rodata_start_addr,16),1)
-    #combine_bin(rodata_bin_name,Icache_flash_bin_name,long(rodata_start_addr,16),0)
+    combine_bin(rodata_bin_name,flash_bin_name,rodata_start_addr,1)
+
 
     # write checksum header
     sum_size = os.path.getsize(flash_bin_name) + 1
@@ -272,22 +297,45 @@ def gen_appbin():
     write_file(flash_bin_name,chr(chk_sum & 0xFF)) 
     #write_file(Icache_flash_bin_name,chr(chk_sum & 0xFF))   	  	
   
-    all_bin_crc = getFileCRC(flash_bin_name)
-    # print all_bin_crc
-    if all_bin_crc < 0:
-        all_bin_crc = abs(all_bin_crc) - 1
-    else :
-        all_bin_crc = abs(all_bin_crc) + 1
-    # print all_bin_crc
-    write_file(flash_bin_name,chr((all_bin_crc & 0x000000FF))+chr((all_bin_crc & 0x0000FF00) >> 8)+chr((all_bin_crc & 0x00FF0000) >> 16)+chr((all_bin_crc & 0xFF000000) >> 24))
-    icache_file = open(flash_bin_name, "rb")
-    if int(os.path.getsize(irom1text_bin_name)):
-        Dcache_bin_len = (int(os.path.getsize(irom1text_bin_name)) + 16 +15 ) &(~15)
-        #print Dcache_bin_len
-        icache_file.seek(Dcache_bin_len ,0)
-    open(Icache_flash_bin_name, "wb").write(icache_file.read())
-    cmd = 'rm eagle.app.sym'
-    os.system(cmd)
+    if combine_bin_flag != 0:
+        print os.path.getsize(app_bin_file)
+        print os.path.getsize(flash_bin_name)
+        app_bin = open(app_bin_file, "rb")
+        write_file(flash_bin_name,app_bin.read())
+       
+    if app_cpu_flag != 1:
+        all_bin_crc = getFileCRC(flash_bin_name)
+        # print all_bin_crc
+        if all_bin_crc < 0:
+            all_bin_crc = abs(all_bin_crc) - 1
+        else :
+            all_bin_crc = abs(all_bin_crc) + 1
+        # print all_bin_crc
+        write_file(flash_bin_name,chr((all_bin_crc & 0x000000FF))+chr((all_bin_crc & 0x0000FF00) >> 8)+chr((all_bin_crc & 0x00FF0000) >> 16)+chr((all_bin_crc & 0xFF000000) >> 24))
+        icache_file = open(flash_bin_name, "rb")
+        if int(os.path.getsize(drom0text_bin_name)):
+            Dcache_bin_len = (int(os.path.getsize(drom0text_bin_name)) + 16 +15 ) &(~15)
+            #print Dcache_bin_len
+            icache_file.seek(Dcache_bin_len ,0)
+        open(Icache_flash_bin_name, "wb").write(icache_file.read())
+        cmd = 'rm eagle.app.sym'
+        os.system(cmd)
+        
+        print "Generate related files successully in folder " + sys.argv[6]
+
+        if sys.argv[3] == '0':
+            print "boot.bin------------------>0x00000"
+            if Dcache_data_len:
+                print "drom0.bin----------------->0x04000"
+                print "irom0_flash.bin----------->0x40000"
+            else :
+                print "irom0_flash.bin----------->0x04000"
+            print "user.ota------------------>(used for OTA)"
 
 if __name__=='__main__':
-    gen_appbin()
+    
+    if len(sys.argv) != 7:
+        print 'Usage: gen_appbin.py eagle.app.out'
+        sys.exit(0) 
+
+    gen_appbin(sys.argv[3],sys.argv[4])
